pyproject.toml
[project]
name = "actionable-segmentation-engine"
version = "0.1.0"
description = "ML + LLM-driven customer segmentation with robustness validation"
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "SK Metric Mavens"}
]
keywords = ["segmentation", "machine-learning", "llm", "customer-analytics"]

dependencies = [
    "polars>=1.0.0",  # Fast DataFrame library (Rust-based, replaces pandas)
    "pydantic>=2.0",
    "jinja2>=3.1",    # HTML templating for visual essays
    "pyyaml>=6.0",    # YAML config for narrative overrides
]

[project.optional-dependencies]
dev = [
    "mypy>=1.5",
    "pyright>=1.1.350",
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-benchmark>=4.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.8.0",
    "pre-commit>=3.6",
    "bandit[toml]>=1.7",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "config*"]

# =============================================================================
# MYPY - Static type checking
# =============================================================================
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "sklearn.*",
    "scipy.*",
    "matplotlib.*",
    "seaborn.*",
    "networkx.*",
    "faiss.*",
    "polars.*",
]
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

# =============================================================================
# PYRIGHT - Fast static type checking (complements mypy)
# =============================================================================
[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "strict"
reportMissingImports = true
reportMissingTypeStubs = false
reportUnusedImport = true
reportUnusedVariable = true
reportDuplicateImport = true
include = ["src"]
exclude = ["**/__pycache__", "build", "dist", ".venv"]

# =============================================================================
# PYTEST - Testing
# =============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=html --cov-report=term-missing"
asyncio_mode = "auto"

# =============================================================================
# RUFF - Fast linting and formatting (Rust-based, replaces black/flake8/isort)
# =============================================================================
[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "B",      # flake8-bugbear
    "UP",     # pyupgrade
    "S",      # flake8-bandit (security)
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "T10",    # flake8-debugger
    "ISC",    # flake8-implicit-str-concat
    "PIE",    # flake8-pie
    "PT",     # flake8-pytest-style
    "Q",      # flake8-quotes
    "RSE",    # flake8-raise
    "RET",    # flake8-return
    "SIM",    # flake8-simplify
    "TID",    # flake8-tidy-imports
    "TCH",    # flake8-type-checking
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented code)
    "PL",     # Pylint
    "TRY",    # tryceratops
    "FLY",    # flynt
    "PERF",   # perflint
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # Line too long (handled by formatter)
    "S101",   # Use of assert (OK in tests)
    "PLR0913", # Too many arguments
    "PLR2004", # Magic value comparison
    "TRY003", # Long exception messages
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",   # assert allowed in tests
    "ARG",    # unused arguments (fixtures)
    "PLR2004", # magic values OK in tests
]

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

# =============================================================================
# BANDIT - Security linting
# =============================================================================
[tool.bandit]
exclude_dirs = ["tests", ".venv"]
skips = ["B101"]  # assert_used (OK for validation)
